+++ package.json
@@ -34,6 +34,7 @@
     "@semantic-release/git": "^7.0.18",
     "@semantic-release/github": "^5.5.5",
     "assert": "^2.0.0",
+    "express": "^4.17.1",
     "jest": "^24.8.0",
     "jsdoc": "^3.6.3",
     "puppeteer": "^2.0.0",
